requirements.txt
click==6.6
requests==2.10.0
jmespath==0.9.0
geojson==1.3.3
paho-mqtt==1.2
